+++ plugin-maven/src/main/java/dev/equo/ide/maven/AbstractP2MojoWithCatalog.java
@@ -16,6 +16,7 @@ package dev.equo.ide.maven;
 import dev.equo.ide.Catalog;
 import dev.equo.ide.CatalogDsl;
 import dev.equo.ide.IdeHook;
+import dev.equo.ide.IdeHookAssistAI;
 import dev.equo.ide.IdeHookM2E;
 import dev.equo.ide.WorkspaceInit;
 import dev.equo.solstice.p2.P2Model;
@@ -112,6 +113,36 @@ public abstract class AbstractP2MojoWithCatalog extends AbstractP2Mojo {
 	public static class EGit extends MavenCatalogDsl {
 		public EGit() {
 			super(Catalog.EGIT);
+		}
+	}
+
+	@Parameter private AssistAI assistAI;
+
+	public static class AssistAI extends MavenCatalogDsl {
+		public AssistAI() {
+			super(Catalog.ASSIST_AI);
+		}
+
+		@Parameter(required = false)
+		private String apiKey;
+
+		@Parameter(required = false)
+		private String modelName;
+
+		@Override
+		protected List<IdeHook> ideHooks() {
+			return List.of(new IdeHookAssistAI());
+		}
+
+		@Override
+		protected void processVersionOverrides() {
+			super.processVersionOverrides();
+			if (apiKey != null) {
+				Catalog.ASSIST_AI.apiKey(workspaceInit(), apiKey);
+			}
+			if (modelName != null) {
+				Catalog.ASSIST_AI.modelName(workspaceInit(), modelName);
+			}
 		}
 	}